+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## v2.0.1 -
 
+- Add
+  - Add tray query feature. User can add custom queries to tray menu for quick access
+    ![](https://raw.githubusercontent.com/Wox-launcher/Wox/refs/heads/master/screenshots/tray_query.png)
+
 - Fix
   - [`File Explorer Search`] Fix an issue that file explorer search plugin cannot navigate on open/save dialog